Move feign-client into spring-boot generator
