+++ lib/index.spec.ts
@@ -58,7 +58,6 @@ describe('getGeneratorNamespaces', () => {
   "entities",
   "entity",
   "export-jdl",
-  "feign-client",
   "gatling",
   "generate-blueprint",
   "git",
@@ -109,6 +108,7 @@ describe('getGeneratorNamespaces', () => {
   "spring-boot",
   "spring-boot:bootstrap",
   "spring-boot:cucumber",
+  "spring-boot:feign-client",
   "spring-boot:jwt",
   "spring-boot:oauth2",
   "spring-boot:websocket",
